EEH-2513 - Merged object_set_creator.py into utils/
